// src/app/shared/model/banner.ts
export interface banner {
    name: string
    image: string
    createdTime: string
}
export interface banner_details {
    id: string
    name: string
    image: string
    createdTime: string
}
export interface banner_details_selected {
    id: string
    name: string
    image: string
    createdTime: string
    checked: boolean
}